+++ src/core/teacher/error/create-edit-available-time-description.error.ts
