+++ src/core/teacher/error/create-edit-available-time-end-time.error.ts
