+++ src/core/teacher/interactor/create-edit-available-time.interactor.ts
@@ -2,9 +2,9 @@ import { Presenter } from '../../common/presenter.interface';
 import { AvailableTimeRepository } from '../available-time.repository';
 import { AvailableTimeRequest } from '../dto/available-time.request';
 import { AvailableTimeResponse } from '../dto/available-time.response';
-import { AvailableTimeDescriptionError } from '../error/create-available-time-description.error';
-import { AvailableTimeEndTimeError } from '../error/create-available-time-end-time.error';
-import { AvailableTimeStartTimeError } from '../error/create-available-time-start-time.error';
+import { AvailableTimeDescriptionError } from '../error/create-edit-available-time-description.error';
+import { AvailableTimeEndTimeError } from '../error/create-edit-available-time-end-time.error';
+import { AvailableTimeStartTimeError } from '../error/create-edit-available-time-start-time.error';
 
 export class CreateEditAvailableTimeInteractor {
   constructor(
